Avoid bad multierror format
Current we get horrible errors like:
```
    traffic.go:180: 1 error occurred:
        	* failed calling a->'://fake.service.local?&protocol=tcp&server=127.0.0.1:0/': call failed from a to dns://fake.service.local?&protocol=tcp&server=127.0.0.1 (using dns): timeout while waiting after 20 attempts, 0/3 successes (last error: 1 error occurred:
        	* unexpected dns response: wanted [], got [1.2.3.4]

        )

```

With this change, it looks instead like:
```
    traffic.go:180: failed calling a->'http://b-alt-3:12345/': call failed from a to http://b-alt-3:12345 (using http): timeout while waiting after 5 attempts, 0/3 successes (last error: expected response code 123, got "200")
```

(different test failure, but same concept)

This is done using `multierror.ErrorFormat`. We used this in a few places, but I refactored it out into a common library so
it can be used by all parts of the code.

The alternative to this would be to just completely fork the multierror
package but I avoided doing that for now.
